+ Performance memory usage,

+++ Profiling/Models/Process/ProcessThreadInfoStruct.cs
@@ -14,8 +14,8 @@ namespace ProcessManager.Profiling.Models.Process
     [StructLayout(LayoutKind.Sequential)]
     public struct ProcessThreadInfoStruct
     {
-        public UInt32 priority;
-        public UInt32 tid;
+        public UInt64 priority;
+        public UInt64 tid;
         public UInt64 startAddress;
         public UInt64 cyclesDelta;
     }
